pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eurusd_quant"
version = "0.1.0"
description = "EURUSD M15 Session Range Breakout backtest MVP"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pandas",
  "numpy",
  "pyarrow",
  "pyyaml",
]

[project.optional-dependencies]
dev = ["pytest"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
